+++ backend/app/management/commands/seed_data.py
@@ -36,6 +36,7 @@ class Command(BaseCommand):
         ):
             self.stdout.write(self.style.SUCCESS("Database already seeded"))
             return
-        sync_metadata(resource_id=postgres.id, workspace_id=workspace.id)
-        sync_metadata(resource_id=metabase.id, workspace_id=workspace.id)
+        # TODO uncomment this before merging. its throwing an error now
+        # sync_metadata(resource_id=postgres.id, workspace_id=workspace.id)
+        # sync_metadata(resource_id=metabase.id, workspace_id=workspace.id)
         self.stdout.write(self.style.SUCCESS("Successfully seeded the database"))
